+++ src/components/Swap/Status/StatusDialog.tsx
@@ -2,7 +2,7 @@ import { Trans } from '@lingui/macro'
 import ErrorDialog, { StatusHeader } from 'components/Error/ErrorDialog'
 import EtherscanLink from 'components/EtherscanLink'
 import Rule from 'components/Rule'
-import SwapSummary from 'components/Swap/ReviewSwap'
+import SwapSummary from 'components/Swap/Summary'
 import useInterval from 'hooks/useInterval'
 import { CheckCircle, Clock, Spinner } from 'icons'
 import ms from 'ms.macro'
